// src/enums/enums.ts
export enum NotificationCategory {
  NEWACCOUNT = 'NEWACCOUNT',
  CONTACT = 'CONTACT',
  NOTIFICATION = 'NOTIFICATION',
  STATUSCHANGE = 'STATUSCHANGE',
  LOGIN = 'LOGIN',
  CHAT = 'CHAT',
  MESSAGE = 'MESSAGE',
  OTP = 'OTP',
  STOCK = 'STOCK',
  GENERAL = 'GENERAL',
  BROADCAST = 'BROADCAST',
  BIRTHDAY = 'BIRTHDAY',
  HOLIDAY = 'HOLIDAY',
  ANNIVERSARY = 'ANNIVERSARY',
  PAYMENT = 'PAYMENT',
  INVESTMENT_APPLIED = 'INVESTMENT_APPLIED',
  INVESTMENT_APPROVED = 'INVESTMENT_APPROVED',
  INVESTMENT_REJECTED = 'INVESTMENT_REJECTED',
  INVESTMENT_PAYMENT_RECEIVED = 'INVESTMENT_PAYMENT_RECEIVED',
  INVESTMENT_EXPIRING_SOON = 'INVESTMENT_EXPIRING_SOON',
  INVESTMENT_EXPIRED = 'INVESTMENT_EXPIRED',
  INVESTMENT_SETTLED = 'INVESTMENT_SETTLED',
  INVESTMENT_EXTENDED = 'INVESTMENT_EXTENDED',
  SALE_CREATED = 'SALE_CREATED',
  SALE_APPLICATION = 'SALE_APPLICATION',
  SALE_APPROVED = 'SALE_APPROVED',
  SALE_PAYMENT = 'SALE_PAYMENT',
  SALE_COMPLETED = 'SALE_COMPLETED',
  SALE_APPLIED = 'SALE_APPLIED',
  SALE_PAYMENT_COMPLETED = 'SALE_PAYMENT_COMPLETED',
  SALE_PAYMENT_SUCCESSFUL = 'SALE_PAYMENT_SUCCESSFUL',
  PAYMENT_RECEIVED = 'PAYMENT_RECEIVED',
  PAYMENT_SUCCESSFUL = 'PAYMENT_SUCCESSFUL',
  PAYMENT_COMPLETED = 'PAYMENT_COMPLETED',
  PAYMENT_REVERSED = 'PAYMENT_REVERSED',
  PAYMENT_REFUND = 'PAYMENT_REFUND',
  NEWS = 'NEWS',
  PAYMENT_PLAN_CREATED = 'PAYMENT_PLAN_CREATED',
  QUERY_SENT = 'QUERY_SENT',
  QUERY_ACKNOWLEDGED = 'QUERY_ACKNOWLEDGED',
  QUERY_REVOKED = 'QUERY_REVOKED',
  SUSPENDED = 'SUSPENDED',
  SUSPENSION_REVOKED = 'SUSPENSION_REVOKED',
  SUSPENSION_COMPLETED = 'SUSPENSION_COMPLETED',
  ALLOCATION = 'ALLOCATION',
}

export enum NotificationPriority {
  HIGH = 'HIGH',
  MEDIUM = 'MEDIUM',
  LOW = 'LOW',
}
// ...
export enum PaymentMethod {
  PARTPAYMENT = 'Part Payment',
  FULLPAYMENT = 'Full Payment',
  CASH = 'CASH',
}

export enum PaymentType {
  PROPERTYFEE = 'PROPERTYFEE',
  FACILITYFEE = 'FACILITYFEE',
  WATERFEE = 'WATERFEE',
  ELECTRICITYFEE = 'ELECTRICITYFEE',
  SUPERVISIONFEE = 'SUPERVISIONFEE',
  AUTHORITYFEE = 'AUTHORITYFEE',
  OTHERFEE = 'OTHERFEE',
  INFRASTRUCTURECOST = 'INFRASTRUCTURECOST',
  AGENCYFEE = 'AGENCYFEE',
  GENERAL = 'GENERAL',
}
// ...
export enum StateStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  AVAILABLE = 'AVAILABLE',
  UNAVAILABLE = 'UNAVAILABLE',
  SUSPENDED = 'SUSPENDED',
  BLOCKED = 'BLOCKED',
  TRASHED = 'TRASHED',
  UNSEEN = 'UNSEEN',
  SEEN = 'SEEN',
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  DECLINED = 'DECLINED',
  DELETED = 'DELETED',
  PAID = 'PAID',
  PARTIALLYPAID = 'PARTIALLYPAID',
  UNPAID = 'UNPAID',
  PAYING = 'PAYING',
  REPLIED = 'REPLIED',
  PURCHASED = 'PURCHASED',
  RELEASED = 'RELEASED',
  ALLOCATED = 'ALLOCATED',
  CANCELED = 'CANCELED',
  VALID = 'VALID',
  EXPIRED = 'EXPIRED',
  RESOLVED = 'RESOLVED',
  INPROGRESS = 'INPROGRESS',
  COMPLETED = 'COMPLETED',
  RESERVED = 'RESERVED',
  SOLD = 'SOLD',
  RESCHEDULED = 'RESCHEDULED',
  SETTLED = 'SETTLED',
  SUPPLIED = 'SUPPLIED',
  PARTIALLYSETTLED = 'PARTIALLYSETTLED',
  PENDINGSETTLEMENT = 'PENDINGSETTLEMENT',
  SUCCESS = 'SUCCESS',
  QUEUED = 'QUEUED',
  SENT = 'SENT',
  FAILED = 'FAILED',
  CLOSED = 'CLOSED',
  DELIVERED = 'DELIVERED',
  REJECTED = 'REJECTED',
  DUE = 'DUE',
  CANCELLED = 'CANCELLED',
  ACKNOWLEDGED = 'ACKNOWLEDGED',
  REVOKED = 'REVOKED',
  UPDATED = 'UPDATED',
  OFFERED = 'OFFERED',
}

export enum InvestmentDuration {
  MONTH = 'MONTH',
  QUARTER = 'QUARTER',
  HALF_YEAR = 'HALF_YEAR',
  YEAR = 'YEAR',
}

export enum ActionType {
  UPDATE = 'UPDATE',
  DELETE = 'DELETE',
  PASSWORDCHANGE = 'CHANGEPASSWORD',
  DECREASE = 'DECREASE',
  INCREASE = 'INCREASE',
  CREATE = 'CREATE',
  CANCEL = 'CANCEL',
  SETTLE = 'SETTLE',
  STATUS_UPDATE = "STATUS_UPDATE",
}